[IMP] zen: verses about code quality
A beautiful code is a code that can easily be read, one that use well
defined function, that is well integrated in its environment. On the
other hand, an ugly piece of code is code that seems overly complicated.

A beautiful code is also a piece of code that is more easily read by
other. Readability is important as it makes it easier to maintain the
software on the long run.