[IMP] zen: verse about ambiguity
Beautiful code is code that doesn't leave you guessing. When you
are face to face with beautiful code, it speaks to you, and you
know exactly how it is going to act.

In order to write beautiful code that is empty of ambiguity, you have
to cleanse yourself from the ambiguity within before you engage in the
act of bringing new code into this world.

When you are not sure of something, deal with it in the moment,
and figure it out, don't let the ambiguity accumelate.